Create Balanced Binary Tree
Given a binary tree, determine if it is height-balanced.

For this problem, a height-balanced binary tree is defined as a binary tree in which the depth of the two subtrees of every node never differ by more than 1.

compare the depths of two sub-tree, if it's larger than 1 ,return -1. else return 0.